+++ header_files/calc.h
@@ -1,7 +1,7 @@
 #pragma once
 #include"../header_files/help.h"
 
-double sciCalc();
+int sciCalc();
 double evaluate_postfix(Token output[], int out_count);
 void shunting_yard(const char *input, Token output[], int *out_count);
 int is_right_assoc(char op);
